abstracted usePayoutStatus hook
